Move handler functions to handler module
